requirements.txt
torch>=1.4.0
torchvision>=0.5.0
pyyaml
einops
setuptools==59.5.0
tqdm
pytorch-lightning